MAD-2/Code/backend/requirements.txt
amqp==5.2.0
async-timeout==4.0.3
bcrypt==4.2.0
billiard==4.2.0
blinker==1.8.2
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
dnspython==2.6.1
email_validator==2.2.0
Flask==3.0.3
Flask-Cors==4.0.1
Flask-Principal==0.4.0
Flask-WTF==1.2.1
idna==3.7
importlib_resources==6.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
kombu==5.3.7
MarkupSafe==2.1.5
numpy==2.0.1
pandas==2.2.2
passlib==1.7.4
prompt_toolkit==3.0.47
PyJWT==2.9.0
python-dateutil==2.9.0.post0
pytz==2024.1
redis==5.0.8
six==1.16.0
tzdata==2024.1
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.0.3
WTForms==3.1.2
